requirements y simplificacion de clases
