+++ backend/api/admin.py
@@ -1,9 +1,5 @@
 from django.contrib import admin
-from .models.trailer import Trailer
-from .models.user import User
-from .models.load import Load
-from .models.accespoint import AccessPoint
-from .models.trip import Trip
+from .models import Trailer,User,Load,AccessPoint,Trip
 
 
 admin.site.register(Trailer)
