+++ backend/requirements.txt
@@ -1,2 +1,4 @@
 pip install django
 pip install djangorestframework
+pip install dj-rest-auth
+pip install django-allauth
